# drustvena_mreza/Web App/api/api.yaml
swagger: '2.0'
info:
  title: Bubbles API
  description: Mrzim svoj život.
  version: 1.0.0
produces:
  - application/json
paths:
...
  /sign-up:
    post:
      tags:
        - General
      summary: Registracija
      parameters:
        - name: username
          in: formData
          required: true
          type: string
        - name: email
          in: formData
          required: true
          type: string
        - name: password
          in: formData
          required: true
          type: string
        - name: firstName
          in: formData
          required: false
          type: string
        - name: lastName
          in: formData
          required: false
          type: string
        - name: middleName
          in: formData
          required: false
          type: string
        - name: address
          in: formData
          required: false
          type: string
        - name: city
          in: formData
          required: false
          type: string
        - name: country
          in: formData
          required: false
          type: string
      responses:
        '200':
          description: Kod uspješne registracije.
        default:
          description: Vraća kolekciju pogreški vezane uz poejdino polje
          schema:
            $ref: '#/definitions/ValidationError'
  /sign-in:
    post:
      tags:
        - General
      summary: Prijava
      parameters:
        - name: username
          in: formData
          required: true
          type: string
        - name: password
          in: formData
          required: true
          type: string
      responses:
        '200':
          description: Uspješna prijava
        default:
          description: Poruka vezana uz neuspješnu prijavu.
          schema:
            $ref: '#/definitions/Error'
...
  '/content/comments/{content_id}':
    get:
      tags:
        - Content
      summary: Vraća sve komentare za dani content_id
      parameters:
        - name: content_id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Komentari
          schema:
            type: object
            properties:
              comments:
                type: array
                items:
                  $ref: '#/definitions/ExtendedComment'
  '/content/comment/{content_id}':
    post:
      tags:
        - Content
      summary: Posta komentar na content sa content_id-om
      parameters:
        - name: content_id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Kod uspjesnog komentiranja
  '/content/like/{id}':
    post:
      tags:
        - Content
      summary: Lajka/unlajka(toggle) content sa idem id
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Uspješno lajkanje
  '/content/likes/{id}':
    get:
      tags:
        - Content
      summary: Dohvaća kolekciju korisnika koji su lajkali content sa id-em id
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Uspješno dohvaćanje
          schema:
            type: object
            properties:
              users:
                type: array
                items:
                  type: object
                  properties:
                    first_name:
                      type: string
                    last_name:
                      type: string
                    middle_name:
                      type: string
                    username:
                      type: string
...
definitions:
...
  ExtendedComment:
    type: object
    properties:
      id:
        type: integer
        format: int32
      content_id:
        type: integer
        format: int32
      user_id:
        type: integer
        format: int32
      created_at:
        type: string
      updated_at:
        type: string
      comment:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      username:
        type: string
...
  Error:
    type: object
    properties:
      error:
        type: string
  ValidationError:
    type: object
    properties:
      errors:
        type: array
        items:
          type: string